requirements.txt
hydra-core==1.1.2
mlflow==1.25.1
bokeh==2.2.3
boto3==1.17.110
boto==2.49.0
botocore==1.20.110
catboost==0.26.1
click==7.1.2
flask==1.1.2
gensim==4.0.0
h5py==2.10.0
ipython==7.18.1
ipywidgets==7.6.3
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter_client==6.1.7
jupyter_console==6.2.0
jupyter_core==4.6.3
jupyterlab==2.2.6
jupyterlab_pygments==0.1.2
jupyterlab_server==1.2.0
jupyterlab_widgets==1.0.0
matplotlib==3.3.1
networkx==2.5
nltk==3.5
numpy==1.22.3
pandas==1.1.3
pillow==8.0.0
plotly==5.3.1
pluggy==0.13.1
pytest-cov==3.0.0
pytest==6.2.5
python==3.8.10
pytorch==1.9.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.8.0
seaborn==0.11.0
setuptools==52.0.0
statsmodels==0.12.0
torch==1.9.0
torchaudio==0.9.0
torchsummary==1.5.1
torchvision==0.10.0
tqdm==4.64.0

omegaconf~=2.1.2